db-migrate added

+++ skeema/schemas/AdventureWorks/test_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `TestTable` (
-  `id` int NOT NULL AUTO_INCREMENT,
+  `id` bigint unsigned NOT NULL AUTO_INCREMENT,
   `FirstName` VARCHAR(50) NOT NULL,
   `LastName` VARCHAR(50) NOT NULL,
   PRIMARY KEY (`id`)
